block: Drop power-of-2 atomic write boundary requirement
The current requirement is that any non-zero atomic write boundary is a
power-of-2. Otherwise, we disable atomic writes (for that bdev).

The atomic write hw unit max must be a power-of-2, and any atomic write
size must be equal or less than that value and also a power-of-2.

To ensure alignment, it follows that that the boundary just needs to be a
multiple of the atomic write hw unit max, so relax the boundary
requirement.

Note that the comment in blk_validate_atomic_write_limits() about reducing
atomic write unit max is not really sensible, as that value should never
exceed the boundary (now), so drop that comment.

Signed-off-by: John Garry <john.g.garry@oracle.com>