iomap: Lift blocksize restriction on atomic writes
Filesystems like ext4 can submit writes in multiples of blocksizes.
But we still can't allow the writes to be split into multiple BIOs. Hence
let's check if the iomap_length() is same as iter->len or not.

It is the responsibility of userspace to ensure that a write does not span
mixed unwritten and mapped extents (which would lead to multiple BIOs).

Signed-off-by: Ritesh Harjani (IBM) <ritesh.list@gmail.com>
jpg: tweak commit message
Signed-off-by: John Garry <john.g.garry@oracle.com>