xfs: Prototype CoW fallback for atomic writes
In the case of a mapping not be properly aligned or a discontiguous
mapping, it is not possible to issue an atomic write. This is why we are
limited to atomically writing a single block.

To lift that limit, in the case described, create a CoW fork for the data
to be written and update the mapping once the DIO completes - this
emulates atomic write behaviour.

Signed-off-by: John Garry <john.g.garry@oracle.com>