xfs: Add xfs_file_dio_write_atomic()
Add xfs_file_dio_write_atomic(), for handling of aligned and misaligned
atomic writes.

A misaligned atomic write is when either of the following occurs:
- the start disk block is not aligned wrt atomic write length
- an atomic write spans multiple extents, whether they are mapped,
  unwritten, or unallocated.

Currently only 1x block is allowed to be written atomically, so a
misaligned atomic write cannot occur.

In the situation of a misaligned atomic write, a CoW-based method is used
to issue the atomic write.

For this CoW-based method to be used, it must be first detected that there
is an misaligned atomic write. In this case, the atomic write must be
retried in this CoW mode.

The detect and retry method similar to that in
xfs_file_dio_write_unaligned() is used here.

An exclusive iolock must be taken in this process, to stop racy access to
CoW extents.

Signed-off-by: John Garry <john.g.garry@oracle.com>