xfs: limit atomic write unit according to agsize
An atomic write which uses HW offload needs to be properly aligned to
the disk block on the block device.

Currently we just check that the mapping startblock is aligned for an
atomic write. However, that is just the startblock within the allocation
group, and the allocation group may not be properly aligned to the
underlying block device atomic write alignment.

As such, limit atomic writes to the greatest power-of-2 which fits in an
AG, so that aligning to the startblock will be mean that we are also
aligned to the disk block.

Signed-off-by: John Garry <john.g.garry@oracle.com>

# Conflicts:
#	fs/xfs/xfs_mount.c