scsi: scsi_debug: Remove sdebug_device_access_info
This structure is not used, so delete it.

It was originally intended for supporting checking for atomic writes
overlapping with ongoing reads and writes, but that support never got
added.

sbc-4 r22 section 4.29.3.2 "Performing operations during an atomic write
operation" describes two methods of handling overlapping atomic writes.
Currently the only method supported is for the ongoing read or write to
complete.

In future, the other handling method - to raise CHECK CONDITION error -
can be supported in the driver, but this structure is probably not going
to be used there anyway.

Signed-off-by: John Garry <john.g.garry@oracle.com>