xfs: iomap COW-based atomic write support
In cases of an atomic write covering misaligned or discontiguous disk
blocks, we will use a CoW-based method to issue the atomic write.

So, for that case, return -ENOPROTOOPT to request that the write be issued in
CoW atomic write mode. The dio write path should detect this, similar to
how misaligned regular DIO writes are handled.

For normal REQ_ATOMIC-based mode, when the range which we are atomic
writing to covers a shared data extent, try to allocate a new CoW fork.
However, if we find that what we allocated does not meet atomic write
requirements in terms of length and alignment, then fallback on the
CoW-based mode for the atomic write.

For scenario in which we are required to allocate > 1x block for the write,
fallback on the CoW-based method. This avoids the scenario in which we
would end up allocating multiple extents for write (and so not suitable for
REQ_ATOMIC-based write), and then need to fallback on the CoW-based method.

The CoW-based method will also be used for an atomc write which appends to
file.

Signed-off-by: John Garry <john.g.garry@oracle.com>