xfs: always roll the transaction when reaping rtgroup cow extents
In general, accumulating large chains of deferred log intents is a bad
idea because this is very stressful on the log subsystem and only
necessary to restart an operation during recovery.  Since log recovery
frees all stale CoW extents, we can roll the transaction between extents
in xreap_rgextent_iter because it only ever gets used for CoW extents.

Cc: <stable@vger.kernel.org> # v6.14
Fixes: fd97fe1112088c ("xfs: fix CoW forks for realtime files")
Signed-off-by: "Darrick J. Wong" <djwong@kernel.org>