xfs: only call xfs_setsize_buftarg once per buffer target
It's silly to call xfs_setsize_buftarg from xfs_alloc_buftarg with the
block device LBA size because we don't need to ask the block layer to
validate a geometry number that it provided us.  Instead, set the
preliminary bt_meta_sector* fields to the LBA size in preparation for
reading the primary super.

It's ok to lose the sync_blockdev call at buftarg creation time for the
external log and rt devices because we don't read from them until after
calling xfs_setup_devices.  We do need an explicit sync for the data
device because we read the primary super before calling
xfs_setup_devices.

This will enable the next patch to validate hw atomic write geometry
against the filesystem geometry.

Signed-off-by: Darrick J. Wong <djwong@kernel.org>