scsi: sd: Add atomic write capabilities device attribute
To allow the user to overwrite the values read from the VPD pages, add
atomic write capabilities device attribute.

This will be useful in scenarios when the devices incorrect reports
atomic write parameters in the VPD pages.

We only add support for max, granularity, and alignment, while not adding
support for boundary. Support is not added for boundary, as this is of
little interest for kernel atomic write support.

Note that issuing an atomic write uses WRITE ATOMIC (16) command, so an
atomic write whose parameters are outside the supported values will be
rejected. That's assuming that WRITE ATOMIC (16) has been properly
implemented (if at all).

Signed-off-by: John Garry <john.g.garry@oracle.com>