dm-stripe: set max_hw_sectors limit
It is possible to break atomic writes by having the bottom device
physical_block_size less than the chunk size, like:

# cat /sys/block/sda/queue/atomic_write_unit_max_bytes
65536
# cat /sys/block/sda/queue/physical_block_size
16384
# cat /sys/block/sda/queue/logical_block_size
512
# cat /sys/block/sda/queue/max_hw_sectors_kb
2147483647
# cat /sys/block/sda/queue/minimum_io_size
16384
#
# vgcreate vg00 --dataalignment=16M /dev/sda /dev/sdb
# lvm lvcreate -i 2 -I 8 -L 4M -n striped_logical_volume vg00
#
# cat /sys/block/dm-0/queue/atomic_write_unit_max_bytes
16384
# cat /sys/block/dm-0/queue/physical_block_size
16384
# cat /sys/block/dm-0/queue/logical_block_size
512
# cat /sys/block/dm-0/queue/max_hw_sectors_kb
2147483647
# cat /sys/block/dm-0/queue/minimum_io_size
16384
#

With a chunk size of 8192, and atomic write of size 16384 is not possible.

For this case, stacking the request queue limits means that the top device
minimum_io_size is raised that of the bottom device and so the top device
atomic write unit max is limited by the bottom/top minimum_io_size, and
not actually the chunk size.

Remedy this by setting the dm-stripe device max_hw_sectors limit in
addition to the chunk size, so the atomic write unit max is also limited
to that.

Signed-off-by: John Garry <john.g.garry@oracle.com>