dm-stripe: limit max_hw_sectors to the chunk size
Currently we use min io size as the chunk size when deciding on the limit
of atomic write size.

Using min io size is not reliable, as this may be mutated when stacking
the bottom device limits.

The block stacking limits will rely on max_hw_sectors in future, so set
this value (to the chunk size).

Introduce a flag - DM_TARGET_STRIPED - and check this in
dm_set_device_limits().

Signed-off-by: John Garry <john.g.garry@oracle.com>