nvme: add an opt-in to use AWUPF
As mentioned at [0], many parts of the atomic write specification are
lacking.

For now, there is nothing which we can do in software about the lack of
a dedicated NVMe write atomic command.

As for reading the atomic write limits, it can be assumed that the per-
namespace values are mostly properly specified and supported.

The specification of NAWUPF is quite clear, however the specification of
NABSPF is less clear. The lack of clarity in NABSPF comes from deciding
whether NABSPF applies when NSABP is 0 - it is asssumed that NSABP does
not apply when NSABP is 0.

As for the per-controller AWUPF, how this values applies to shared
namespaces is missing in the specification. Furthermore, the value is in
terms of logical blocks, which is an NS entity.

It would be preferred to stop honouring AWUPF altogther, but this may
disable atomic write support for many devices which only specify AWUPF.

Allow users of such devices to use atomic writes at their own risk by
adding a per-NVMe subsystem sysfs file to enable reading of AWUPF.

A udev rule like the following could be used to auto-enable:
SUBSYSTEM=="nvme-subsystem", ACTION=="add", RUN+="/bin/sh -c 'echo 1 > /sys/$devpath/use_awupf'"

Having a per-controller file could also work as an alternative, but it
is inlikely that a system will have a mix of controllers which we would
want to selectively enable reading AWUPF.

Note that AWUPF not allow effects atomic write support, but also the
physical block size reported for the device.

[0] https://lore.kernel.org/linux-nvme/20250707141834.GA30198@lst.de/

Signed-off-by: John Garry <john.g.garry@oracle.com>