nvme: do not use AWUPF
As mentioned at [0], many parts of the atomic write specification are
quite lacking.

For now, there is nothing which we can do in software about the lack of
a dedicated NVMe write atomic command (apart from the checks in
nvme_setup_rw() to ensure that we are not issued a request which falls
outside the limits).

As for reading the atomic write limits, it can be assumed that the per-
namespace values are mostly properly specified and supported.

The specification of NAWUPF is quite clear, however the specification of
NABSPF is less clear. The lack of clarity in NABSPF comes from deciding
whether NABSPF applies when NSABP is 0 - it is asssumed that NSABP does
not apply when NSABP is 0.

We cannot always trust that the value is reported in AWUPF is correct,
so just stop using it altogether. If a device only reporrts values in
AWUPF, atomic writes will just not be supported.

This is qute an unfortunate approach, but it may push NVMe firmware
developers to always provide per-NS atomic pararms.

[0] https://lore.kernel.org/linux-nvme/20250707141834.GA30198@lst.de/

Signed-off-by: John Garry <john.g.garry@oracle.com>