md: disable atomic writes if any bad blocks found
RAID1 and RAID10 personalities support bad blocks. In handling read or
writing from a region which has bad blocks, the read and write bio is split
as to not access the bad block sectors.

Splitting of writes is not allowed for atomic writes. As such, bad blocks
are incompatible with atomic writes. If an atomic write ever occurs over
sectors which has bad blocks, then the write is rejected.

It is not ever expected in practice that we will ever see devices which
support atomic writes and also require bad block kernel support. However
it is not proper to say that atomic writes are supported for a device
which actually has bad blocks.

md maintains a persistent feature flag to say whether bad block management
is active on the array, MD_FEATURE_BAD_BLOCKS.

If this is set, then disable atomic writes for the array. In addition, if
bad blocks are found, the atomic writes are asynchronously disabled.

Signed-off-by: John Garry <john.g.garry@oracle.com>