scsi: scsi_debug: Clear sd_dp->defer_t in sdebug_q_cmd_complete()
When a scsi_cmnd is recycled, it may check the old value, so clear it.

Such a situation is when we issue a fake timeout, as sd_dp->defer_t is
not reset in the queue_rq() path.

It would be nicer to have such code in sdebug_init_cmd_priv(), but that
is only called once per request init (and not per cycle).

Signed-off-by: John Garry <john.g.garry@oracle.com>